requirements.txt
streamlit
langchain
langchain-community
langchain-chroma
langchain-huggingface
chromadb
beautifulsoup4
lxml
tiktoken
sentence-transformers
huggingface_hub
openai
langchain-google-genai
python-dotenv
